moved GistDetail into own folder, splits container/component
